+++ index.js
@@ -1,2 +1,7 @@
-require('coffee-script');
+CoffeeScript = require('coffee-script');
+
+if (typeof(CoffeeScript.register) == 'function') {
+  CoffeeScript.register();
+}
+
 module.exports = require('./src/hamlc');